+++ src/main/java/com/mobble/mobbleserver/domain/chat/room/ClubRoomInfo.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.domain.chat.room;
 
-import com.mobble.mobbleserver.domain.club.core.Club;
+import com.mobble.mobbleserver.domain.club.Club;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
