+++ src/main/java/com/mobble/mobbleserver/domain/clubMember/ClubMember.java
@@ -1,7 +1,7 @@
 package com.mobble.mobbleserver.domain.clubMember;
 
 import com.mobble.mobbleserver.domain.article.ArticleType;
-import com.mobble.mobbleserver.domain.club.core.Club;
+import com.mobble.mobbleserver.domain.club.Club;
 import com.mobble.mobbleserver.domain.common.BaseEntity;
 import com.mobble.mobbleserver.domain.member.Member;
 import jakarta.persistence.*;
